Fix 5.8 Mass includes; first successful build of both modules
FMassEntityHandle must be included explicitly via Mass/EntityHandle.h
(MassCore) - it is not reliably visible through MassEntityTypes.h in a
fresh TU despite appearing in its include list. ExecutionFlags cast fixed
to uint8 (5.8 narrowed the field). red_hopeEditor Development: clean link,
32 s.

+++ Source/RedHopeSim/Private/RHWanderProcessor.cpp
@@ -9,7 +9,7 @@
 URHWanderProcessor::URHWanderProcessor()
 	: EntityQuery(*this)
 {
-	ExecutionFlags = (int32)EProcessorExecutionFlags::All;
+	ExecutionFlags = (uint8)EProcessorExecutionFlags::All;
 	bRequiresGameThreadExecution = true; // reads world subsystems (clock, sim world)
 }
